More-like core build system

+++ pulsar_modules/methods/mbe/MBE.cpp
@@ -1,13 +1,9 @@
 
 #include<pulsar/modulebase/SystemFragmenter.hpp>
 #include<pulsar/util/IterTools.hpp>
+#include<pulsar/math/Cast.hpp>
 #include "pulsar_modules/methods/mbe/MBE.hpp"
 #include "pulsar_modules/methods/method_helpers/MethodHelpers.hpp"
-
-/*
- * TODO once computations cache results, compute interactions by
- *   "Re-running" the computations and then assembling derivatives
- */
 
 //Options
 const std::string method_key="METHOD_KEY";
@@ -20,7 +16,7 @@ DerivReturnType MBE::deriv_(size_t Order,const Wavefunction& Wfn){
     const System& Mol=*Wfn.system;
     NMerSetType NMers=
             create_child_from_option<SystemFragmenter>(system_key)->fragmentize(Mol);
-    std::vector<double> Cs,Result(std::pow(3*Mol.size(),Order));
+    std::vector<double> Cs,Result(numeric_cast<size_t>(std::pow(3*Mol.size(),Order)));
     std::vector<Wavefunction> Wfns;
     for(const typename NMerSetType::value_type& NMerI : NMers){
         Cs.push_back(NMerI.second.weight);
